:recycle: Extract execute.ts
